+++ src/store/modules/attendance.js
@@ -6,13 +6,14 @@ const state = {
   data: [],
   status: {},
   update: {},
+  loading: Boolean
 };
 
 const actions = {
   async getAttendanceCustom({commit}){
-    console.log('get custom attendance')
+    // console.log('get custom attendance')
     const res = await httpCommons.get(apiName+'/customGetAttendance');
-    console.log(res)
+    // console.log(res)
     commit("SET_GET_DATA_ATTENDANCE", res.data);
   },
   
@@ -25,7 +26,7 @@ const actions = {
       var yyyy = today.getFullYear();
       const today1 = today
       var todayDate = yyyy + "-" + mm + "-" + dd;
-      console.log(todayDate)
+      // console.log(todayDate)
     //   // today = "18/11/2022"
     //   // console.log(today);
     //   date = today;
@@ -47,13 +48,13 @@ const actions = {
     params.append("filter", "created_at||cont||"+todayDate)
     
     const res = await httpCommons.get(apiName, {params : params});
-    console.log(res)
+    // console.log(res)
     commit("SET_GET_DATA_ATTENDANCE", res.data);
   },
 
   async actionGetAllAttendenceByFilter({ commit }, param) {
     const res = await httpCommons.get(apiName, {params : param});
-    console.log(res);
+    // console.log(res);
     commit("SET_GET_DATA_ATTENDANCE", res.data);
   },
 
@@ -75,7 +76,8 @@ const actions = {
   },
 
   async saveBulkAttendance({ commit, dispatch }, data) {
-    console.log(data);
+    // console.log(data);
+    commit("SET_LOADING_ATTENDANCE", true);
     try {
       const res = await httpCommons.post(apiName + "/bulk", data);
       const result = {
@@ -83,32 +85,34 @@ const actions = {
         actions: res.status,
         data: res.data,
       };
-      console.log("res : " + res);
-      console.log(result);
+      // console.log("res : " + res);
+      // console.log(result);
       commit("SET_CHECK_ATTENDANCE", result);
-      dispatch("actionGetAllAttendence");
+      dispatch("getAttendanceCustom");
+      
     } catch (error) {
-      console.log(error);
+      // console.log(error);
       const result = {
         status: "duplicate",
         actions: 201,
       };
       commit("SET_CHECK_ATTENDANCE", result);
     }
+    commit("SET_LOADING_ATTENDANCE", false);
   },
 
   async updateOvertime({ commit, dispatch }, data) {
     try {
       const res = await httpCommons.patch(apiName + `/${data.id}`, data);
-      console.log(res);
+      // console.log(res);
       const result = {
         status: res.statusText,
         actions: res.status,
         data: res.data,
       };
-      console.log("overtime_update : " + result);
+      // console.log(result);
       commit("SET_UPDATE_ATTENDANCE", result);
-      dispatch("actionGetAllAttendence");
+      // dispatch("actionGetAllAttendence");
     } catch (error) {
       const result = {
         status: "duplicate",
@@ -119,7 +123,7 @@ const actions = {
   },
 
   async checkAttendance({ commit, dispatch }, data) {
-    console.log(data);
+    // console.log(data);
     try {
       const res = await httpCommons.post(apiName + "/checkAttendance", data);
       const result = {
@@ -127,7 +131,7 @@ const actions = {
         actions: res.status,
         data: res.data,
       };
-      console.log(result.data);
+      // console.log(result.data);
       commit("SET_CHECK_ATTENDANCE", result);
     } catch (error) {
       const result = {
@@ -141,13 +145,13 @@ const actions = {
   async deleteAttendanceById({ commit, dispatch }, id) {
     try {
       const res = await httpCommons.delete(apiName + "/" + id);
-      console.log("res " + res);
+      // console.log("res " + res);
       const result = {
         status: res.statusText,
         actions: res.status,
       };
       commit("SET_DELETE_ATTENDANCE", result);
-      dispatch("actionGetAllAttendence");
+      // dispatch("actionGetAllAttendence");
     } catch (error) { }
   },
 };
@@ -159,6 +163,10 @@ const mutations = {
 
   SET_CHECK_ATTENDANCE(state, status) {
     state.status = status;
+  },
+
+  SET_LOADING_ATTENDANCE(state, loading) {
+    state.loading = loading;
   },
 
   SET_UPDATE_ATTENDANCE(state, status) {
@@ -174,6 +182,7 @@ const getters = {
   getDataAllAttendance: (state) => state.data,
   getStatusAttendance: (state) => state.status,
   getBulkAttendance: (state) => state.status,
+  getLoadingAttendance: (state) => state.loading,
 };
 
 export default {