+++ CHANGELOG.md
@@ -98,3 +98,15 @@
     - validasi jam lembur nya ijo& merah nya jangan yg cetar, warna kalem
 - Modul Attendance Bulanan
     - On going
+
+# 0.0.1 (Update 06/02/2023)
+- Modul Attendance Mingguan (feedback MJ)
+    - Notifikasi Duplikasi data
+    - Add component overlay loading
+
+# 0.0.1 (Update 07/02/2023)
+- Modul Attendance Mingguan (feedback MJ)
+    - Filter Attendance
+    - Tambah, Hapus, dan Edit kehadiran by filter
+- Modul Pinjaman (feedback MJ)
+    - Mengganti get data employee